Changes to the Message wire format:
-requestId has become message id, and message id has been moved up as a field of
Message, as opposed to Request
-a new type of Message, called Failure, has been added. The Failure message is
used to indicate that a request failed on the server side. Now that this message
type has been added, we can now report faliures back over the wire, and throw
exceptions on the client side

NOTE: Server request failures result in RuntimeExceptions in
ViewerServiceClient, which means that the DevMode server will shut down.

Review by: mmendez

git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@6762 8db76d5a-ed1c-0410-87a9-c151d255dfc7