Update some old selection script templates for modern dev mode
Copied the definition of isHostedMode() in from IFrameLinker's template, which is really the most up to date. Without this change, trying to run hosted mode on some of these linkers erroneously gives the "this module needs to be recompiled" message instead of the better "this linker doesn't support hosted mode".

http://gwt-code-reviews.appspot.com/376801/show
Review by: jat


git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@7957 8db76d5a-ed1c-0410-87a9-c151d255dfc7