Replacing Cell#consumesEvents() with Cell#getConsumedEvents(), and using the return value to sink only the required events for each Cell Widget. https://wave.google.com/wave/#restored:wave:googlewave.com%252Fw%252BDbKnvP3NA
Adds a new method Cell#handlesSelection() to indicate that the cell updates the selection model, and that the container widget should not update the selection model on mouse events. This patch also moves selection before firing the event to the cell, as the Cell may depend on selection.
https://wave.google.com/wave/#restored:wave:googlewave.com%252Fw%252BUB51B1NjA

Review at http://gwt-code-reviews.appspot.com/707803


git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@8420 8db76d5a-ed1c-0410-87a9-c151d255dfc7