Makes the terminology used by AbstractListViewAdapter and AbstractPager consistent with HasData.  HasData#setRowValues() is now HasData#setRowData().  Also renames ListViewAdapter to ListDataProvider so it is more obvious that the class provides data to a Cell widget.  The term "View" has been replaced by "Display" (for pagers) and "DataDisplay" (for data providers).  "View" insinuates an MVC design, but in our case the widgets do have some logic, so we chose the more generic term "Display" to describe them.  This patch looks huge, but 95% of it is a bunch of renames.
This change also adds a new method HasData#setVisiableRangeAndClearData(), which can be used to simulataneously set a new range, clear all data in the Cell widget, and fire a RangeChangeEvent.  This is useful for tables that show search results. For example, when a user types a search query, you can call setVisibleRangeAndClearData() to go back to the first page, clear old search results, and fire a RangeChangeEvent to trigger a new data request for new search results.

Review at http://gwt-code-reviews.appspot.com/772801


git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@8553 8db76d5a-ed1c-0410-87a9-c151d255dfc7