Rollback of 847801
*** Reason for rollback ***

Broke users of setHTML() with null parameter, where the intended method is ambiguous.

*** Original change description ***

Add SafeHtml support to UI widgets (2)

This is a first-pass at adding SafeHtml support to a subset of the widgets. In cases where the class implements HasHTML, HasSafeHtml has been implemented as well. In constructors that accept a string that can be parsed as Html, a SafeHtml constructor has been added. In cases where a method accepts a string that can be parsed as Html, a corresponding SafeHtml method has been added.

Review at http://gwt-code-reviews.appspot.com/847801


git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@8747 8db76d5a-ed1c-0410-87a9-c151d255dfc7