Replaces the public and obnoxious String EntityProxy#getId() with the private ProxyImpl#encodedId(), which is set by th server and opaque to the client.
EntityProxyChange is now id based, rather than record based.
I tried to make the corresponding fixes to samples/expenses,
but it's broken at the moment. (Rather than fixing it to
work with String ids, it might be better to expend the same
amount of effort making Long ids work again.)

EntityProxyId is now parameterized to the type of its
EntityProxy, so that the find request doesn't suck.

Review at http://gwt-code-reviews.appspot.com/902801


git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@8830 8db76d5a-ed1c-0410-87a9-c151d255dfc7