Fix issue 5462 where DynaTableRf explodes when favoriting a new person.
This fix hides the favorite checkbox by default, and shows it if the person has been persisted. This is a pretty dirty hack, so please review it to make sure it's not too dirty. The correct way to fix this issue would be to re-work DynaTableRf so that the favorite boolean is a property on the person, but this is a workaround until that can be done.

Review at http://gwt-code-reviews.appspot.com/1043802


git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@9133 8db76d5a-ed1c-0410-87a9-c151d255dfc7