Reduces size of the compile report (soyc) for large projects with many split points by:
- Using a common intern dictionary for methodDependency
reports from the same report.
- Sorting the interned dictionary by number of
references (so the most referenced strings have single
digit indices).
- Interning strings in the leftoversStatus report.
- Using single letter function names.

These changes amount to a further 20 to 30 percent reduction
in the size of the report for two large projects I've tried.

This also corrects a problem in the splitStatus and
leftoversStatus reports. The list of classes used
to construct the report was not sorted.

Review at http://gwt-code-reviews.appspot.com/1171801

Review by: kprobst@google.com

git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@9332 8db76d5a-ed1c-0410-87a9-c151d255dfc7