Public(t.broyer@gmail.com):
Adds a unit-test to explicitly check that redundant calls to
showActivityWidget are okay. The test is needed because
AcceptsOneWidget passed to the Activity by ActivityManager is not
directly the one you passed to its setDisplay() method, so
ActivityManager could (theoretically) evolve in the future and break
this use case, which isn't documented as being a "valid use case".

Also "fixes" what looks like an oversight in AsyncActivity#start.

Review by rjrjr@google.com
http://gwt-code-reviews.appspot.com/999802

Review by: robertvawter@google.com

git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@9583 8db76d5a-ed1c-0410-87a9-c151d255dfc7