Adding table rendering tests to micro benchmarks. Table rendering tests are multiple orders of magnitude slower than the existing basic tests, so I seperated them into a seperate mirco benchmark.  I modified WidgetCreation so it could be used for both instead of copying udles of code.  You can see the app running at http://jlabanca-testing.appspot.com/.
I also made some changes to improve test accuracy.  I found that the tests get slower as elements are attached to the DOM and as the test runs, so instead of running all iterations of a single test, we run all tests one iteration at a time.  We also alternate the order of the tests, so the last test added isn't always the last test to run.  Finally, we remove each rendered widget after we add it and record the time, as the existence of a widget could affect the timing of the next test. This seems to lead to more reasonable numbers in practice.

Review at http://gwt-code-reviews.appspot.com/1394802

Review by: rjrjr@google.com

git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@9925 8db76d5a-ed1c-0410-87a9-c151d255dfc7