Make AttachableComposite delegate setStyleName to the wrapped Widget.
Review at http://gwt-code-reviews.appspot.com/1433802

Lack of this is making templates that set the style of AttachableComposite
widgets
fail, because when setStyleName is called the element hasn't been set yet. It's
clear at this point that we need underlying support for this, instead of
replicating this kind of logic all over the place. Again, I'm only overriding
setStyleName because that's what UiBinder uses and that's what's holding us back
a bit. But this will probably be priority number one when we're ironing out the
API/implementation.

Review by: rjrjr@google.com

git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@10152 8db76d5a-ed1c-0410-87a9-c151d255dfc7