Rollback of r10697 due to broken builds
*** Original change description ***

Change default of UiBinder.useLazyWidgetBuilders to true, in
preparation for deleting the old code.

Also fixes unreported bad code gen for things like <a href="{uh.oh}"> when
safehtml is off. Fix is a bit squirrelly, but the intent is to delete the code
in question almost immediately if nothing blows up with new default.


git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@10711 8db76d5a-ed1c-0410-87a9-c151d255dfc7