Fixing a bug in CellBrowser where using a shared SelectionModel across all lists and using KeyboardSelectionPolicy.BOUND_TO_SELECTION causes the lists to fight for selection.  The result is that items in lower index lists (to the left) end up being selected instead of items in the highest index list (rightmost).  This is a fairly complicated problem that requires custom handling of selection state within CellBrowser.
Also fixing a bug in HasDataPresenter where changes to the temporary state aren't propagated to the new pending state if a selection event interupts the SelectionModel part of the state resolution loop.  The problem is that we update the new pending state as we go along, but in most cases the new pending state won't exist until we're halfway through the SelectionModel resolution code.  The fix is to copy all modified fields from the temporary state into the new pending state at the end of the selection resolution logic, when we know the new pending state exists.

Review at http://gwt-code-reviews.appspot.com/1585803

Review by: skybrian@google.com

git-svn-id: http://google-web-toolkit.googlecode.com/svn/trunk@10731 8db76d5a-ed1c-0410-87a9-c151d255dfc7